requirements_up.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
exceptiongroup==1.3.0
fake-useragent==1.5.1
frozenlist==1.7.0
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.56.1
greenlet==3.2.3
gspread==6.1.2
gspread-dataframe==4.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
outcome==1.3.0.post0
pandas==2.2.3
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
telebot==0.0.5
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
